requirements.txt
anyio==3.7.0
apt-clone==0.2.1
attrs==23.1.0
autoflake==2.1.1
bcc==0.12.0
black==23.3.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Pillow==9.5.0
tk==0.1.0
